update commit and undo

+++ src/undo.sh
@@ -1,24 +1,24 @@
 #!/bin/bash
-case $1 in
-  add )
-    #undo ADD
-    git reset --keep
-    ;;
-  commit )
-    #undo commit
-    git reset HEAD~1 --mixed
-    ;;
-  push )
-    #undo PUSH
-    # as written is this blog post: http://megakemp.com/2016/08/25/git-undo/
-    git reset --hard $(git rev-parse --abbrev-ref HEAD)@{${1-1}};
-    ;;
-  * )
-    echo "git gud undo: $1 is not a valid command. See 'git gud man'."
-    ;;
-esac
-
-
-
-
+if [ ! $# == 1 ]; then
+  echo "usage: git gud undo [add|commit|push]"
+else
+  case $1 in
+    add )
+      #undo ADD
+      git reset --keep
+      ;;
+    commit )
+      #undo commit
+      git reset HEAD~1 --mixed
+      ;;
+    push )
+      #undo PUSH
+      # as written is this blog post: http://megakemp.com/2016/08/25/git-undo/
+      git reset --hard $(git rev-parse --abbrev-ref HEAD)@{${1-1}};
+      ;;
+    * )
+      echo "git gud undo: $1 is not a valid command. See 'git gud man'."
+      ;;
+  esac
+fi
